# docs/api/openapi.json
{
  "components": {
    "schemas": {
...
      "GeoPolygon": {
        "properties": {
          "coordinates": {
            "items": {
              "items": {
                "maxItems": 2,
                "minItems": 2,
                "prefixItems": [
                  {
                    "type": "number"
                  },
                  {
                    "type": "number"
                  }
                ],
                "type": "array"
              },
              "type": "array"
            },
            "title": "Coordinates",
            "type": "array"
          },
          "type": {
            "const": "Polygon",
            "default": "Polygon",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "coordinates"
        ],
        "title": "GeoPolygon",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "Page_Scene_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/Scene"
            },
            "title": "Items",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "required": [
          "items"
        ],
        "title": "Page[Scene]",
        "type": "object"
      },
...
      "ProblemDetail": {
        "description": "RFC 7807. Served with media type application/problem+json \u2014 see\nrouters/errors.py, which is the only place that constructs these.",
        "properties": {
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detail"
          },
          "instance": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Instance"
          },
          "status": {
            "title": "Status",
            "type": "integer"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "type": {
            "default": "about:blank",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "title",
          "status"
        ],
        "title": "ProblemDetail",
        "type": "object"
      },
      "Scene": {
        "description": "Not pinned by a literal example in the contract doc (only the\nendpoint list is). Kept intentionally small and aligned with the\nprovenance fields the StatusStrip component and the evidence dossier\nboth need \u2014 see docs/DESIGN_SYSTEM.md \u00a76.5.",
        "properties": {
          "acquired_utc": {
            "title": "Acquired Utc",
            "type": "string"
          },
          "footprint": {
            "$ref": "#/components/schemas/GeoPolygon"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "product_hash": {
            "title": "Product Hash",
            "type": "string"
          },
          "sensor": {
            "title": "Sensor",
            "type": "string"
          },
          "status": {
            "enum": [
              "ingested",
              "preprocessing",
              "analysed"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "id",
          "sensor",
          "acquired_utc",
          "footprint",
          "product_hash",
          "status"
        ],
        "title": "Scene",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "SAGARDRISHTI API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/scenes": {
      "get": {
        "operationId": "list_scenes_api_v1_scenes_get",
        "parameters": [
          {
            "description": "minlon,minlat,maxlon,maxlat",
            "in": "query",
            "name": "bbox",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "minlon,minlat,maxlon,maxlat",
              "title": "Bbox"
            }
          },
          {
            "in": "query",
            "name": "start",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Start"
            }
          },
          {
            "in": "query",
            "name": "end",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "End"
            }
          },
          {
            "in": "query",
            "name": "sensor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Sensor"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 500,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_Scene_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Scenes",
        "tags": [
          "scenes"
        ]
      }
    },
    "/api/v1/scenes/{scene_id}": {
      "get": {
        "operationId": "get_scene_api_v1_scenes__scene_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "scene_id",
            "required": true,
            "schema": {
              "title": "Scene Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Scene"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "No such resource. problem+json, with `detail` naming the id."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Scene",
        "tags": [
          "scenes"
        ]
      }
    },
    "/api/v1/scenes/{scene_id}/analyse": {
      "post": {
        "operationId": "analyse_scene_api_v1_scenes__scene_id__analyse_post",
        "parameters": [
          {
            "in": "path",
            "name": "scene_id",
            "required": true,
            "schema": {
              "title": "Scene Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Accepted \u2014 a job id is returned; poll /jobs/{id}."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "No such resource. problem+json, with `detail` naming the id."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "501": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "The endpoint is in the contract but its pipeline is not built. An expected state today, not a failure \u2014 clients branch on it."
          }
        },
        "summary": "Analyse Scene",
        "tags": [
          "scenes"
        ]
      }
    },
...
  }
}